b2b_sales_enablement/appointment_webhook/requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0
